created a new module infobip-spring-data-jdbc-annotation-processor-common used as base for creating custom annotation processors with different naming strategies
