+++ infobip-spring-data-jdbc-annotation-processor-common/pom.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>com.infobip</groupId>
+        <artifactId>infobip-spring-data-querydsl</artifactId>
+        <version>5.2.1-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>infobip-spring-data-jdbc-annotation-processor-common</artifactId>
+
+    <properties>
+        <!-- DEPENDENCY VERSIONS -->
+        <auto-service.version>1.0-rc6</auto-service.version>
+        <compile-testing.version>0.19</compile-testing.version>
+    </properties>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.google.auto.service</groupId>
+            <artifactId>auto-service</artifactId>
+            <version>${auto-service.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.querydsl</groupId>
+            <artifactId>querydsl-apt</artifactId>
+            <version>${querydsl.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.querydsl</groupId>
+            <artifactId>querydsl-sql-codegen</artifactId>
+            <version>${querydsl.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.data</groupId>
+            <artifactId>spring-data-relational</artifactId>
+        </dependency>
+    </dependencies>
+</project>
